modularized utils functions
