+++ packages/foam-core/src/utils/hashtags.ts
@@ -1,13 +1,14 @@
-import { isSome } from './index';
+import { isSome } from './core';
+import matchAll from 'string.prototype.matchall';
 
-const HASHTAG_REGEX = /(^|\W|\/)#([\w_-]+\b)/gm;
-const WORD_REGEX = /(^|\W|\/)([\w_-]+\b)/gm;
+const HASHTAG_REGEX = /(^|[ ])#([\w_-]+\b)/gm;
+const WORD_REGEX = /(^|[ ])([\w_-]+\b)/gm;
 
 export const extractHashtags = (text: string): Set<string> => {
   return isSome(text)
     ? new Set(
-        Array.from(text.matchAll(HASHTAG_REGEX), m => m[2])
-          .map(tag => tag.trim())
+        Array.from(matchAll(text, HASHTAG_REGEX))
+          .map(m => m[2].trim())
           .filter(tag => !isNumeric(tag))
       )
     : new Set();
@@ -16,7 +17,7 @@ export const extractHashtags = (text: string): Set<string> => {
 export const extractTagsFromProp = (prop: string | string[]): Set<string> => {
   const text = Array.isArray(prop) ? prop.join(' ') : prop;
   return isSome(text)
-    ? new Set(Array.from(text.matchAll(WORD_REGEX), m => m[2]))
+    ? new Set(Array.from(matchAll(text, WORD_REGEX)).map(m => m[2].trim()))
     : new Set();
 };
